Reorganizes built-in libs
