+++ libs/index.js
@@ -1,0 +1,15 @@
+/**=============================================================================
+ *
+ * Copyright (c) 2013 - 2018 Seth Purcell
+ * Licensed under Apache License v2.0 with Runtime Library Exception
+ *
+ * See LICENSE.txt in the project root for license information.
+ *
+ =============================================================================*/
+
+"use strict";
+
+module.exports = {
+
+    String: require("./LoString")
+};
