--- testing/linking/LoadAndGoTest.js
@@ -1,24 +1,0 @@
-/**
- * Created by spurcell on 5/28/16.
- */
-
-const LoadAndGo = require('../../linker/LoadAndGo');
-const LMS = require('../../linker/LocalModuleSpace');
-const util = require('util');
-
-module.exports['basics'] = {
-
-    "success": function (test) {
-
-        // test.expect(3);
-
-        var lms = new LMS(__dirname + '/../programs');
-        var program = new LoadAndGo(lms, 'factorial');
-
-        program.run([10]).then(function () {
-
-            test.done();
-        }
-        ).done();
-    }
-};
